requirements/pyproject.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements/pyproject.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
cffi==1.16.0
    # via cryptography
cryptography==42.0.8
    # via pyjwt
ecdsa==0.19.0
    # via python-jose
fastapi-slim==0.111.0
idna==3.7
    # via anyio
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via
    #   fastapi-slim
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.2
pyjwt==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
    # via pydantic-settings
python-jose==3.3.0
pytz==2024.1
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi-slim
typing-extensions==4.12.2
    # via
    #   fastapi-slim
    #   pydantic
    #   pydantic-core
